[yum] update based build scripts.

+++ packages/yum/build-rpm.sh
@@ -3,9 +3,13 @@
 LANG=C
 
 PACKAGE=$(cat /tmp/build-package)
-USER_NAME=$(cat /tmp/build-user)
 VERSION=$(cat /tmp/build-version)
+SOURCE_BASE_NAME=$(cat /tmp/build-source-base-name)
+USER_NAME=$(cat /tmp/build-user)
 DEPENDED_PACKAGES=$(cat /tmp/depended-packages)
+USE_RPMFORGE=$(cat /tmp/build-use-rpmforge)
+USE_ATRPMS=$(cat /tmp/build-use-atrpms)
+BUILD_OPTIONS=$(cat /tmp/build-options)
 BUILD_SCRIPT=/tmp/build-${PACKAGE}.sh
 
 run()
@@ -17,18 +21,17 @@ run()
     fi
 }
 
+if ! id $USER_NAME >/dev/null 2>&1; then
+    run useradd -m $USER_NAME
+fi
+
+yum_options=
 distribution=$(cut -d ' ' -f 1 /etc/redhat-release | tr 'A-Z' 'a-z')
-case $distribution in
-    fedora)
-	distribution_version=$(cut -d ' ' -f 3 /etc/redhat-release)
-	;;
-    centos)
-	distribution_version=$(cut -d ' ' -f 4 /etc/redhat-release)
-	;;
-esac
-
-rpmbuild_options=""
-
+if grep -q Linux /etc/redhat-release; then
+    distribution_version=$(cut -d ' ' -f 4 /etc/redhat-release)
+else
+    distribution_version=$(cut -d ' ' -f 3 /etc/redhat-release)
+fi
 if ! rpm -q ${distribution}-release > /dev/null 2>&1; then
     packages_dir=/var/cache/yum/core/packages
     release_rpm=${distribution}-release-${distribution_version}-*.rpm
@@ -36,9 +39,46 @@ if ! rpm -q ${distribution}-release > /dev/null 2>&1; then
     run rpm -Uvh --force ${packages_dir}/ca-certificates-*.rpm
 fi
 
+if test "$USE_RPMFORGE" = "yes"; then
+    if ! rpm -q rpmforge-release > /dev/null 2>&1; then
+	architecture=$(cut -d '-' -f 1 /etc/rpm/platform)
+	rpmforge_url=http://packages.sw.be/rpmforge-release
+	rpmforge_rpm_base=rpmforge-release-0.5.2-2.el5.rf.${architecture}.rpm
+	wget $rpmforge_url/$rpmforge_rpm_base
+	run rpm -Uvh $rpmforge_rpm_base
+	rm $rpmforge_rpm_base
+	sed -i'' -e 's/enabled = 1/enabled = 0/g' /etc/yum.repos.d/rpmforge.repo
+    fi
+    yum_options="$yum_options --enablerepo=rpmforge"
+fi
+
+if test "$USE_ATRPMS" = "yes"; then
+    case "$(cat /etc/redhat-release)" in
+	CentOS*)
+	    repository_label=CentOS
+	    repository_prefix=el
+	    ;;
+	*)
+	    repository_label=Fedora
+	    repository_prefix=f
+	    ;;
+    esac
+    cat <<EOF > /etc/yum.repos.d/atrpms.repo
+[atrpms]
+name=${repository_label} \$releasever - \$basearch - ATrpms
+baseurl=http://dl.atrpms.net/${repository_prefix}\$releasever-\$basearch/atrpms/stable
+gpgkey=http://ATrpms.net/RPM-GPG-KEY.atrpms
+gpgcheck=1
+enabled=0
+EOF
+    yum_options="$yum_options --enablerepo=atrpms"
+fi
+
 if ! rpm -q groonga-repository > /dev/null 2>&1; then
     run rpm -Uvh http://packages.groonga.org/${distribution}/groonga-repository-1.0.0-0.noarch.rpm
 fi
+
+rpmbuild_options="${BUILD_OPTIONS}"
 
 case $distribution in
     fedora)
@@ -56,9 +96,9 @@ case $distribution in
 	;;
 esac
 
-run yum update -y
-run yum install -y rpm-build tar ${DEPENDED_PACKAGES}
-run yum clean packages
+run yum update ${yum_options} -y
+run yum install ${yum_options} -y rpm-build tar ${DEPENDED_PACKAGES}
+run yum clean ${yum_options} packages
 
 if ! id $USER_NAME >/dev/null 2>&1; then
     run useradd -m $USER_NAME
@@ -83,12 +123,28 @@ mkdir -p rpm/BUILD
 mkdir -p rpm/RPMS
 mkdir -p rpm/SRPMS
 
-cp /tmp/mroonga-$VERSION.tar.gz rpm/SOURCES/
-cp /tmp/${PACKAGE}.spec rpm/SPECS/
+if test -f /tmp/${SOURCE_BASE_NAME}-$VERSION-*.src.rpm; then
+    if ! rpm -Uvh /tmp/${SOURCE_BASE_NAME}-$VERSION-*.src.rpm; then
+        cd rpm/SOURCES
+        rpm2cpio /tmp/${SOURCE_BASE_NAME}-$VERSION-*.src.rpm | cpio -id
+        if ! yum info tcp_wrappers-devel >/dev/null 2>&1; then
+            sed -i'' -e 's/tcp_wrappers-devel/tcp_wrappers/g' ${PACKAGE}.spec
+        fi
+        if ! yum info libdb-devel >/dev/null 2>&1; then
+            sed -i'' -e 's/libdb-devel/db4-devel/g' ${PACKAGE}.spec
+        fi
+        sed -i'' -e 's/BuildArch: noarch//g' ${PACKAGE}.spec
+        mv ${PACKAGE}.spec ../SPECS/
+        cd
+    fi
+else
+    cp /tmp/${SOURCE_BASE_NAME}-$VERSION.* rpm/SOURCES/
+    cp /tmp/${PACKAGE}.spec rpm/SPECS/
+fi
 
 chmod o+rx . rpm rpm/RPMS rpm/SRPMS
 
-rpmbuild -ba $rpmbuild_options rpm/SPECS/${PACKAGE}.spec
+rpmbuild -ba ${rpmbuild_options} rpm/SPECS/${PACKAGE}.spec
 EOF
 
 run chmod +x $BUILD_SCRIPT
